% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/blat_install.R
\name{check_install}
\alias{check_install}
\title{Function to check whether Blat is installed correctly.}
\usage{
check_install()
}
\value{
logical.
}
\description{
Function to check whether Blat is installed correctly.
}

