% Generated by roxygen2 (4.0.2): do not edit by hand
\name{install_blat}
\alias{install_blat}
\title{Function to install the Blat software}
\usage{
install_blat(arch = 64, force = FALSE)
}
\arguments{
\item{arch}{Architechture. Either 32 or 64.}

\item{force}{logical. Force installation or ask for confirmation first?}
}
\description{
Function to install the Blat software
}

