% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{blatr}
\alias{blat}
\alias{blatr}
\alias{blatr-package}
\title{A simple wrapper for Blat (for Windows) SMTP Mailer in R.}
\usage{
blat(...)
}
\arguments{
\item{...}{arguments to blat}
}
\description{
Blat is a command line email tool for Windows with many features. This
package is a wrapper for using Blat from within R. For information about
Blat, see \code{http://www.blat.net}, for the Blat license, see
\code{http://www.blat.net/?docs/license.txt}.
}
\details{
To use the \code{blatr} package, Blat needs to be installed. This can be done
manually, by downloading the executables from the website \code{www.blat.net}
and place them in the \code{blatr} installation directory. You can also use
\code{blatr:::install_blat} to let \code{blatr} do it for you. You can always
check that the needed files are placed correctly using
\code{blatr:::check_install}.

To get started, see the documentation for \code{blat}.
}
\examples{
\dontrun{
   # With attachment
   blat(f      = "Your Name <your@email.com>",
        to     = "your-recipients@email.com",
        s      = "The subject",
        server = "server.address",
        attach = "C:\\\\path\\\\to\\\\attachment.txt",
        body   = "The text you wish to send.")

   # With file as body
   blat(f        = "Your Name <your@email.com>",
        to       = "your-recipients@email.com",
        s        = "The subject",
        server   = "server.address",
        attach   = "C:\\\\path\\\\to\\\\attachment.txt",
        filename = "C:\\\\path\\\\to\\\\file\\\\with\\\\body.txt")

}
}
\author{
Stefan Milton Bache
Send an email using blat for Windows

This is a wrapper which calls Blat to send emails. For documentation
on the options, see http://www.blat.net/syntax/syntax.html.
You should not use "-" in the argument names, this is done for you.
Simply use name=value pairs as arguments.

Blat can use a file as body of the email, in which case this is
ordinarily the first argument in the Blat command. However,
blatr uses the named argument \code{filename} to specify this.
Unnamed arguments do not work with blatr.
}

